Added layout for myProjects page
