+++ Client/src/Components/Header/header.css
@@ -5,11 +5,14 @@
 		height: auto;
 		color: white;
 	}
+	hr.seperationLine{
+		margin: 0 !important;
+	}
 	#headerContainer p{
 		width: 33vw;
 	}
 	#headerContainer p:hover{
-		font-size: 3.5vw;
+		font-size: 1.2em;
 	}
 	.nav-links{
 		width: 33vw;
