+++ Client/src/Components/MyProjects/myProjects.css
@@ -1,0 +1,184 @@
+@media only screen and (min-width: 501px){
+    #myProjectContainer{
+        color: white;
+        width: 100%;
+        height: 90vh;
+    }
+    #clickableLinks{
+        display: flex;
+        justify-content: center;
+    }
+    a#trelloImage{
+        width: 4%;
+        height: auto;
+    }
+    #myProjectContainer .card{
+        background-color: black;
+        width: 100%;
+        height: 100%;
+        padding-left: 3vw;
+        padding-right: 3vw;
+        padding-top: 4vw
+    }
+    #myProjectContainer .card-reveal{
+        background-color: slategray;
+    }
+
+    #trelloImage img{
+        margin-left: 20vw;
+        padding-right: 15vw;
+        height: 2vh;
+    }
+    div.card-content{
+        color: white;
+        display: inline;
+        
+    }
+
+    .card-title{
+        color: white !important;
+        font-size: 4vw !important;
+    }
+    .card-reveal p{
+        color: white;
+        font-size: 3vw;
+    }
+    #myProjectContainer .seperationLine{
+
+    }
+
+
+}
+
+@media only screen and (max-width: 500px){
+    #websiteUrl{
+        background-color: black;
+        color: white;
+        font-size: 20px;
+        margin: 0;
+    }
+    #clockSpecialty{
+        width: 22vh;
+    }
+    #trelloImage img{
+        width: 22vw;
+        padding: .3em;
+        margin: .1em;
+    }
+    #trelloImage:hover{
+        width: 25vw;
+        background-color: darkgreen;
+        border-bottom: darkgreen solid .2em;
+    }
+    /* Arrow for show more about the site */
+    i.material-icons.right{
+        margin-left: -2em !important;
+        margin-top: .2em !important;
+    }
+    /* Up arrow */
+    i.material-icons.right:hover{
+        color: darkgreen;
+    }
+
+    li.slide.selected{
+        padding-top: 2em;
+    }
+
+    div .card{
+        z-index: 0;
+        height: 100vh;
+        color: white;
+        background-color: black;
+        padding: 2em;
+    }
+    /* live Demo button */
+    div .customButton{
+
+        /* appearance: button;
+        -webkit-appearance: button;
+        -moz-appearance: button; */
+
+        display: inline-block;
+        font-size: 5vw;
+        padding: .2em;
+
+        border: none;
+        border-radius: 4px;
+        -moz-border-radius: 4px;
+        -webkit-border-radius: 4px;
+        
+        
+        background-color: #282c34;
+        text-decoration: none;
+        text-align: center;
+        color: white;
+    }
+    div .customButton:hover{
+        font-size: 5.5vw;
+        background-color: darkgreen;
+
+    }
+
+    .waves-block{
+        height: 46vh;
+        margin-bottom: 2em;
+    }
+    p{
+        font-size: 20px;
+        justify-content: center;
+    }
+
+    .card-reveal p{
+        color: black;
+    }
+    .card .card-image{
+        border-bottom: darkgreen solid .1em;
+        text-align: -webkit-center;
+    }
+
+    .card .card-title{
+        font-size: 22px;
+        padding-left: 20px;
+    }
+    .card .card-title.activator{
+        text-align: -webkit-center;
+        text-align: -moz-center;
+    }
+
+    .card p a{
+        font-size: 20px;
+    }
+
+    .card .card-content{
+        color: white;
+        background-color: black;
+        padding-top: 1em;
+        height: 44vh;
+    }
+    /* Github Image */
+    .card .card-content p{
+        width: 13vw;
+        height: auto;
+        padding: .3em;
+        margin-top: .3em;
+        margin-bottom: .3em;
+    }
+    .card .card-content p:hover{
+        width: 14vw;
+        background-color: darkgreen;
+        border-radius: 50%;
+
+    }
+
+    .carousel .control-next.control-arrow:before:hover{
+        border-left: 8px solid darkgreen;
+    }
+
+    /* Carousel and slider-wrapper allow duv to takes up the whole screen instead of 400px */
+    .carousel{
+        height: auto;
+    }
+    .slider-wrapper{
+        height: 90vh !important;
+    }
+}
